Move BuildStep to an interface and add BuildStepImpl with the gory details
